Refactor of the ERXNavigation package out of ERExtensions to its own project.
