+++ Applications/BugTracker/Sources/er/bugtracker/Factory.java
@@ -25,17 +25,16 @@ import com.webobjects.foundation.NSKeyValueCoding;
 
 import er.bugtracker.pages.ReportPage;
 import er.directtoweb.ERD2WFactory;
-import er.directtoweb.delegates.ERDObjectWasCreatedDelegate;
 import er.directtoweb.interfaces.ERDQueryPageInterface;
 import er.directtoweb.pages.ERD2WInspectPage;
 import er.directtoweb.pages.ERD2WQueryPage;
 import er.extensions.appserver.ERXSession;
-import er.extensions.appserver.navigation.ERXNavigationManager;
 import er.extensions.eof.EOEnterpriseObjectClazz;
 import er.extensions.eof.ERXEC;
 import er.extensions.eof.qualifiers.ERXPrimaryKeyListQualifier;
 import er.extensions.foundation.ERXStringUtilities;
 import er.extensions.localization.ERXLocalizer;
+import er.navigation.ERXNavigationManager;
 
 /**
  * Central page creation class. All workflow things should go here or in the super classes. The session
