+++ Applications/BugTracker/Sources/er/bugtracker/components/PageWrapper.java
@@ -15,8 +15,8 @@ import com.webobjects.directtoweb.D2WPage;
 import com.webobjects.foundation.NSKeyValueCoding;
 import com.webobjects.foundation.NSMutableDictionary;
 
-import er.extensions.appserver.navigation.ERXNavigationManager;
-import er.extensions.appserver.navigation.ERXNavigationState;
+import er.navigation.ERXNavigationManager;
+import er.navigation.ERXNavigationState;
 import er.extensions.components._private.ERXSubmitButton;
 import er.extensions.foundation.ERXPatcher;
