+++ Examples/D2W/ERModernMoviesDemo/Sources/er/modern/movies/demo/Application.java
@@ -1,7 +1,7 @@
 package er.modern.movies.demo;
 
 import er.extensions.appserver.ERXApplication;
-import er.extensions.appserver.navigation.ERXNavigationManager;
+import er.navigation.ERXNavigationManager;
 
 public class Application extends ERXApplication {
 	public static void main(String[] argv) {
