+++ Examples/D2W/ERMovies/.classpath
@@ -22,5 +22,6 @@
 	<classpathentry kind="con" path="WOFramework/ERTaggable"/>
 	<classpathentry kind="con" path="WOFramework/ERAttachment"/>
 	<classpathentry kind="con" path="WOFramework/ERPrototypes"/>
+	<classpathentry kind="con" path="WOFramework/ERXNavigation"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
