+++ Frameworks/Core/ERXNavigation/.classpath
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<classpath>
+	<classpathentry kind="src" path="Sources"/>
+  <classpathentry combineaccessrules="false" kind="con" path="WOFramework/ERExtensions"/>
+  <classpathentry combineaccessrules="false" kind="con" path="WOFramework/ERJars"/>
+  <classpathentry combineaccessrules="false" kind="con" path="WOFramework/ERPrototypes"/>
+  <classpathentry combineaccessrules="false" kind="con" path="WOFramework/JavaWOExtensions"/>
+  <classpathentry combineaccessrules="false" kind="con" path="WOFramework/WOOgnl"/>
+  <classpathentry combineaccessrules="false" kind="con" path="WOFramework/JavaEOAccess"/>
+  <classpathentry combineaccessrules="false" kind="con" path="WOFramework/JavaEOControl"/>
+  <classpathentry combineaccessrules="false" kind="con" path="WOFramework/JavaFoundation"/>
+  <classpathentry combineaccessrules="false" kind="con" path="WOFramework/JavaJDBCAdaptor"/>
+  <classpathentry combineaccessrules="false" kind="con" path="WOFramework/JavaWebObjects"/>
+  <classpathentry combineaccessrules="false" kind="con" path="WOFramework/JavaXML"/>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
+	<classpathentry kind="output" path="bin"/>
+</classpath>
