+++ Frameworks/Core/ERXNavigation/.settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,5 @@
+#Mon Aug 20 15:56:35 PDT 2012
+eclipse.preferences.version=1
+encoding//Components/Nonlocalized.lproj/ERXNavigationMenu.wo/ERXNavigationMenu.wod=UTF-8
+encoding//Components/Nonlocalized.lproj/ERXNavigationMenuItem.wo/ERXNavigationMenuItem.wod=UTF-8
+encoding/Components=UTF-8
