+++ Frameworks/Core/ERXNavigation/Components/ERXNavigationMenuItem.api
