+++ Frameworks/Core/ERXNavigation/Sources/er/navigation/ERXModernNavigationMenu.java
@@ -1,4 +1,4 @@
-package er.extensions.appserver.navigation;
+package er.navigation;
 
 import org.apache.log4j.Logger;
