+++ Frameworks/Core/ERXNavigation/build.properties
@@ -1,0 +1,14 @@
+classes.dir = bin
+project.name=ERXNavigation
+project.name.lowercase=erxnavigation
+project.type=framework
+principalClass = 
+customInfoPListContent =
+eoAdaptorClassName =
+cfBundleVersion =
+cfBundleShortVersion =
+cfBundleID = org.mywoapp
+javaVersion = 1.5+
+component.wellFormedTemplateRequired=false
+component.inlineBindingPrefix=$
+component.inlineBindingSuffix=
