+++ Frameworks/Core/ERXNavigation/woproject/classes.exclude.patternset
@@ -1,0 +1,1 @@
+build.properties
