+++ Frameworks/Core/ERXNavigation/woproject/classes.include.patternset
@@ -1,0 +1,2 @@
+**/*.class
+*.properties
