+++ Frameworks/Core/ERXNavigation/woproject/resources.exclude.patternset
@@ -1,0 +1,3 @@
+**/*.eomodeld~/
+**/*.woa/**
+**/*.framework/**
