+++ Frameworks/Core/ERXNavigation/woproject/resources.include.patternset
@@ -1,0 +1,3 @@
+Components/**/*.wo/**/*
+Components/**/*.api
+Resources/**/*
