+++ Frameworks/Core/ERXNavigation/woproject/wsresources.exclude.patternset
@@ -1,0 +1,3 @@
+**/*.woa/**
+**/*.framework/**
+**/*.eomodeld~/**
