+++ Frameworks/Core/ERXNavigation/woproject/wsresources.include.patternset
@@ -1,0 +1,1 @@
+WebServerResources/**/*
